requirements.txt
aiohttp
gidgethub
gitpython